Rename redis.js to backend-redis.js
